requirements-api.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
sentence-transformers==3.1.1
faiss-cpu==1.8.0
python-dotenv==1.0.1
requests==2.32.3
numpy==1.26.4
pydantic==2.9.2
python-multipart==0.0.9
